edited header and moved header into our authprovide container
