+++ package.json
@@ -13,6 +13,7 @@
     "react-dom": "^18.1.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "react-toastify": "^9.0.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
