+++ src/Components/Header.js
@@ -1,15 +1,19 @@
 import {Link} from 'react-router-dom'
+import { useContext } from 'react'
+import AuthContext from '../context/AuthContext'
 const Header = () =>{
- 
+    let {user, logoutUser} = useContext(AuthContext)
     return(
         <>
             <nav className='navbar'>
-                <Link to ='/'>Login</Link>
+                {user ? (<button onClick={logoutUser}>Logout</button>):<Link to ='/'>Login</Link>}
                 <span> | </span>
                 <Link to ='/home'>Home</Link>
                 <span> | </span>
                 <Link to ='/register'>Register</Link>
 
+               
+                
             </nav>
         </>
     )
