+++ src/Components/Main.js
@@ -1,6 +1,6 @@
 import { useEffect,useState } from "react"
 import {Routes, Route} from 'react-router-dom'
-import Login from "./Login"
+import Login from "../Pages/Login"
 // import User from '../Pages/User'
 import PrivateRoute from '../utils/PrivateRoute'
 import { AuthProvider } from '../context/AuthContext'
@@ -8,12 +8,14 @@ import ProtectedPage from '../Pages/User'
 import Test from '../Pages/Test'
 import Register from '../Pages/Register'
 import Home from '../Pages/Home'
+import Header from "./Header"
 
 
 const Main = ()=>{
 
     return(
         <AuthProvider>
+            <Header/>
             <Routes>
                 <Route element ={<PrivateRoute/>}>
                     <Route path ='/home' element ={<Home/>}/>
